hg: add flake8 config file
Summary:
flake8 looks very useful, though not all of its default rules apply to the Mercurial codebase.

I've added the start of a config files so we can keep it but start to disable or tweak rules that don't fit the existing code
style. I added a few to get started.

Reviewed By: DurhamG

Differential Revision: D6791079

fbshipit-source-id: c98662f57d21ed8e4f5def465b3c92765dd69f92