update eden to target the treemanifest source of truth
Summary:
Now that we've moved the hg source around, update eden
to point to the source of truth rather than the synced subdir.

I spoke with andrewjcg about the use of header_namespace here;
it is required to be able to reference the headers without changing
the code under scm/hg to be relative to the fbcode root.  We can't
change that code in that way because it has to build standalone
for the opensource consumers.

Remove the duplicated code while we're in here to make sure
that we're not accidentally referencing it anywhere.

Reviewed By: simpkins

Differential Revision: D6792969

fbshipit-source-id: ec55e98c0bde2488c213b1fd9f8baae9868a350c