mononoke: move changed_entry_stream to manifest_utils
Summary:
It will be used in other places, such as repoconfig.
Unfortunately I had to split tests and implementation. The reason is that tests
depend on blobrepo. Blobrepo crate depends on mercurial_types and that causes
circular dependency. It shouldn't be the case, because
mercurial_types_unittests should be a separate crate, so there may be a bug in
the build system. For now I've split the tests and implementation.

Reviewed By: jsgf

Differential Revision: D6857098

fbshipit-source-id: 1c0af234564dd5628ba6d63d28cf54d7a2dc9217