hgproto: make the handler return future of response and future of unprocessed data
Summary:
The client sends us a request that we parse and return response for it. As simple as that.

When the client sends `unbundle` request we need to return an acknowledgment that we are ready to read a bundle2 stream from it.
This means that we first parse the request, then send the response, then parse the stream.

This will be changed in future diffs so that we might:
Parse request from client -> [send acknowledgments -> parse data from client]* -> send response;

Reviewed By: farnz

Differential Revision: D6833723

fbshipit-source-id: 0c13b1e3253bd664bc565eacb5b58a9fc53d4efa