packaging: add PKGBUILD for building on Arch Linux
Summary:
Backport the build script so it's now buildable on Linux.

Add a hard-coded fallback version `4.4.2` in `setup.py` when `hg` cannot be
found. This makes the build not depending on hg - to allow bootstrapping on
Arch Linux - you can scp the files needed to do the build, without requiring
a special hg that can read fbsource.

Note: an hg that can work with the LFS extension is still needed. But that
is just `sudo pacman -S mercurial`.

Reviewed By: DurhamG

Differential Revision: D6827349

fbshipit-source-id: 45ea75175d8890a845d987cfe2be33251cd1a806