hg: remove assert from treemanifest code
Summary:
This assert appears to have always been wrong, but only respected when
built by clang. The SubtreeIterator class is created with an empty string in the
path member, and the only entry to the class is the next function which
immediately asserts that path is not empty. Since the only place path is set is
later in the function, there's no way it could start off as anything other than
empty.

Reviewed By: quark-zju

Differential Revision: D6891356

fbshipit-source-id: c55e616461d4a29308ca6e22be51ea02962d9df3