hg: make simplecache work with Windows FS
Summary:
When simplecache is configured to use `local` cache, it tries to write files
into a cache directory on a machine. The file names are the key names and they
can contain ':' characters, which are prohibited in Windows filenames. Let's
replace them with '!' which are equivalently unlikely to be parts of actual
files in the repo, yet are allowed on Windows.

Reviewed By: ryanmce

Differential Revision: D6900184

fbshipit-source-id: d901f2ff6e0e6e8b2a0ca93db32b4a836dccd048