lfs: allow global cache to be disabled
Summary:
Previously there is no way to disable the global cache, since an empty
`lfs.usercache` means a default location.

This patch changes the meaning of an empty `lfs.usercache` to "disable".

This is useful for cases when the user cannot write to `~/.cache` and is
using the default configuration.

Reviewed By: ryanmce

Differential Revision: D6890047

fbshipit-source-id: 352dea5142178d5331e37a48957d9d2b8af26ab5