p4fastimport: allow sync commits only when imported repository is up-to-date
Summary:
Previously we allowed a sync commit to happen at any time, but imported the
full repository to bring it into sync. This is slow for very large repositories
and might be accidentally triggered. We change this behavior to only do a sync
commit if only the imported client changed but nothing else.

Reviewed By: dsp

Differential Revision: D6838370

fbshipit-source-id: eecdfb0ea295585058784f8d1f70de5f8c733645