mononoke: add pylz4 crate
Summary:
Add a separate crate that uses lz4 in the same way as python lz4 library. The
main difference is that first 4 bytes are length of the raw data in le32
format. The reason for moving it in a separate crate is to use pylz4 for
remotefilelog getfiles method.

Also removed one panic and replaced it with error.

Reviewed By: jsgf

Differential Revision: D6884918

fbshipit-source-id: 1b05381c045a1f138ab28820175289233b07a91d