hg: fix unbounded manifest memory usage during pushrebase
Summary:
During pushrebase we graft each commit one by one onto the destination
commit. To do so we iterate over the list of changectx's in variable `revs` and called
rev.manifest() on each one. Because `revs` is a list, and because each
rev.manifest() stores a reference to the manifest in the changectx, we end up
keeping every manifest alive which uses a lot of memory.

To fix this, let's treat revs as mutable pop it as we go.

Reviewed By: ryanmce

Differential Revision: D6912485

fbshipit-source-id: 18d1df470dfbc52c215104e620c9a6471ff32571