phabstatus: catch TypeErrors thrown from revision lookup code
Summary:
In T25434376, we see that this data can in some cases be None, and when
we index into None, we get a TypeError:

```
'NoneType' object has no attribute '__getitem__'
```

I re-arranged the except statements and added TypeError to the second one.
I'm not totally happy with this since it's a programming error not actually
well dealt with, but this is the quickest fix. Future commits will fix things
more thoroughly.

Reviewed By: DurhamG

Differential Revision: D6826446

fbshipit-source-id: a21abba25817b746cb77a1f652ed15a189fce8fe