flake8: enable part of flake8 checks
Summary:
Previously the linter is disabled. It's still nice to have it detect obvious
errors. So let's enable a subset of the warnings that won't complain the
current codebase.

I have put an `#` before warnings that could be useful but are incompatible
with the current code base. We can fix them later.

hgsubversion tests are excluded explicitly for now to reduce the files that
need cleanup.

Reviewed By: ryanmce

Differential Revision: D6934532

fbshipit-source-id: b2b29139cb3cb3c70a9cdeb9dc29074a7217a828