progress: add spinner context manager for easy spinners
Summary:
Some actions can take a long time, and not all of them have progress bars.
When this happen the user is left with an empty prompt and doesn't know that
Mercurial is actually working.

This adds a context manager which can be used to give long-running processes a
spinner-style progress bar, indicating that something is happening in the
background.

Reviewed By: farnz

Differential Revision: D6937079

fbshipit-source-id: 396d7b11157b4f13e0c12dac1e2403d9269938ab