fixcorrupt: improve usability
Summary:
We have seen several cases that 00changelog.i got filled with 0s at the end
after a hard reboot. Improve fixcorrupt tool to be able to detect and fix
that. Previously it will not treat node=nullid as corrupted, and it would
also trust linkrev=0, which will truncate the entire changelog.

Reviewed By: ryanmce

Differential Revision: D6952921

fbshipit-source-id: b77e85f23c304ebde3740d6ed70effa1bc792928