mononoke: add wirepack converter
Summary:
Wirepack packer code has an interesting logic that can be generalized.
For example, it can be used on the push side to convert b2xtreegroup wirepack
stream into stream on entries that can be used by Commit Api. This logic will
be implemented in the next diffs.

In this diff we separate wirepack logic into
two parts: converter stream, that converts wirepack stream into another stream,
and a trait that is used by converter.

Reviewed By: lukaspiatkowski

Differential Revision: D6965830

fbshipit-source-id: 726a84ba44c7f85b5aac48e702755dc904c9c6ab