mercurial-bundles: replace the flat Bundle2Item with better structured one
Summary:
The previous Bundle2Stream is a stream of enum that can represent any part of bundle2 - being it a header, inner payload or end marker.

This commit changes the Bundle2Stream to be a stream of more structured elements, where each element is a header with corresponding payload attached.

This new approach is easier to handle, because the way a bundle2 stream is supposed to be dealt with is running an operation per header (where the type of operation depends on type of header). With the old Bundle2Stream the "handling" of bundle2 required basically reparsing it's content to the structures that the new design returns directly.

Reviewed By: jsgf

Differential Revision: D6965529

fbshipit-source-id: 03f7cf57fec5c38addf32bdef3ca56d3239288b9