radixbuf: avoid using unstable features in buck build
Summary:
`test::Bencher` is an unstable feature, which is enabled by 3rd-party crate
`rustc-test`. However, `rustc-test` does not work with buck build. So let's
workaround that by allowing all usage of `test::Bencher` to be disabled by a
feature. And turn on that feature in buck build. Cargo build will remain
unchanged.

Reviewed By: singhsrb

Differential Revision: D7011703

fbshipit-source-id: e08ba9516bf7fadb6edb52ab107e0172df0aaf5b