remotenames: do not over-check for svn/git push targets
Summary:
Previously, if the user had an svn or git push target set up, we would
bail from the normal push --to process even if the users were pushing to a
different location. This is silly. Instead, let's only bail if the user is
actually pushing to an svn or git destination repository.

This bug was discovered while testing the updated svn-to-hg sync scripts.

Differential Revision: D7023173

fbshipit-source-id: bec5e069bc043c2f73fb334bdaeb18a6fad93eef