hg: make pushrebase and infinitepush tree part creation consistent
Summary:
The treemanifest extension now uses a pattern of
treemod._cansendtrees() to detect if it should send trees. pushrebase and
infinitepush were using a different pattern where they were sending trees if
they existed (in infinitepush's case it was ignoring the sendtrees config
entirely). Let's make these consistent.

The new behavior requires that trees exist for the commits we're trying to send.
In order for this to work, we'll need the logic that automatically converts a
flat manifest into a treemanifest so that we're never missing a tree we want to
send.

Reviewed By: markbt

Differential Revision: D7004982

fbshipit-source-id: 7dd6788173a0393d7a129123d5d62256d64b8a64