restack: don't run rebase within a transaction
Summary:
Per discussion with @[100000771202578:kulshrax] [1], this isn't necessary anymore. This complicates the code, and prevents `amend --fixup` from using IMM, so let's remove it.

[1]
```
Haha, so at first I was like "wtf why does it do that"
then I remembered that inhibit used to be a thing
I'm pretty sure it does that to cause inhibit's post-transaction hook to fire in the right spot
which is now irrelevant since inhibit is gone
I can't really think of a reason why that transaction needs to be there (unless there's something else going on that's subtle)
```

Reviewed By: quark-zju

Differential Revision: D7031379

fbshipit-source-id: 73fdeec845c61af7e3bbaf17c7f42c8ab70588c2